RM-7014 add code to check an rm site has been created before running the patches that require an rm site

+++ rm-community/rm-community-repo/source/java/org/alfresco/module/org_alfresco_module_rm/patch/v32/RMv32HoldChildAssocPatch.java
@@ -107,15 +107,18 @@ public class RMv32HoldChildAssocPatch extends AbstractModulePatch
     @Override
     public void applyInternal()
     {
-        qnameDAO.updateQName(ASSOC_FROZEN_RECORDS, ASSOC_FROZEN_CONTENT);
-        for (NodeRef filePlan : filePlanService.getFilePlans())
+        if(qnameDAO.getQName(ASSOC_FROZEN_RECORDS) != null)
         {
-            for (NodeRef hold : holdService.getHolds(filePlan))
+            qnameDAO.updateQName(ASSOC_FROZEN_RECORDS, ASSOC_FROZEN_CONTENT);
+            for (NodeRef filePlan : filePlanService.getFilePlans())
             {
-                for (ChildAssociationRef ref : nodeService.getChildAssocs(hold, ASSOC_FROZEN_CONTENT, ASSOC_FROZEN_RECORDS))
+                for (NodeRef hold : holdService.getHolds(filePlan))
                 {
-                    holdService.removeFromHold(hold, ref.getChildRef());
-                    holdService.addToHold(hold, ref.getChildRef());
+                    for (ChildAssociationRef ref : nodeService.getChildAssocs(hold, ASSOC_FROZEN_CONTENT, ASSOC_FROZEN_RECORDS))
+                    {
+                        holdService.removeFromHold(hold, ref.getChildRef());
+                        holdService.addToHold(hold, ref.getChildRef());
+                    }
                 }
             }
         }
